removed some warnings and unused variables
the ExposedFormat warning fix is super janky

Also -- just noticed a bug where the Play button is disabled (clicking it does nothing) when you import a WAV file, but once you press the space bar, the play button is engaged and clickable again. Not sure how long that bug was there.