Now everything in Effect returns false
Even though nothing related to the EffectManager was used in Project, a bunch of stuff needed to be removed from Effect before I could remove the line to read the EffectManager.h file in Project.h. This commit looks really sketchy to me and I wouldn't be surprised if I reverted it very soon.