Lame temporary fix for SnapTo dropdown
There must be some array that has a # of elements hard-coded in it, and if I can decrement that # by 1, I should be able to remove the SnapTo dropdown. Until then, I managed to fix the alignment by putting a dummy element where the SnapTo dropdown used to be. Call it the Indiana Jones fix.